requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3; python_version < "3.11"
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6; platform_system == "Windows"
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
gdown==5.1.0
grpcio==1.62.1
huggingface-hub==0.22.2
hydra-core==1.3.0
idna==3.6
imageio==2.34.0
jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy-loader==0.4
lightning-utilities==0.11.2
markdown==3.6
markupsafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-cupti-cu12==12.1.105; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-nvrtc-cu12==12.1.105; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-runtime-cu12==12.1.105; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cudnn-cu12==8.9.2.26; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cufft-cu12==11.0.2.54; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-curand-cu12==10.3.2.106; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusolver-cu12==11.4.5.107; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusparse-cu12==12.1.0.106; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nccl-cu12==2.18.1; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nvjitlink-cu12==12.4.127; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nvtx-cu12==12.1.105; platform_system == "Linux" and platform_machine == "x86_64"
omegaconf==2.3.0
open-metric-learning==2.1.7
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.3.0
protobuf==4.25.3
pyparsing==3.1.2
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.1.2
pytz==2024.1
pyyaml==6.0.1
qudida==0.0.4
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
setuptools==69.2.0
six==1.16.0
soupsieve==2.5
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tifffile==2024.2.12
tokenizers==0.15.2
torch==2.1.1
torchmetrics==1.3.2
torchvision==0.16.1
tqdm==4.66.2
transformers==4.39.3
triton==2.1.0; platform_system == "Linux" and platform_machine == "x86_64"
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.0
werkzeug==3.0.2
yarl==1.9.4
